Minimum distance for friendlies set!
Also added a way to adjust their speed in the inspector. Note that the
lower the speed number is, the faster they go - so a speed of 1 is the
original, way too fast speed.

Also, I finally name a branch something other than minDistance to figure
something else, and that's the branch that has the minimum distance fix.
Whatever...