Third person cam and some other fixes
Establishes a movable third-person camera. I also made an adjustment to
the “HP” script and “PlayerHealth” script.

The “HP” script created a public game object called “Enemy” that was
never assigned, so I assigned the “EnemySphere” prefab to it in the
inspector.

 “PlayerHealth” was using the “PlayerController” script when it should
have been using “PlayerControllerPhysics”; on that note, I also deleted
“PlayerController”).