Main Menu - bare bones draft
This is a rough draft of the main menu. It is made in a new scene, and
required me to change the build settings of the project in order to
index scenes.

It consists of a screen with two buttons; “Play Game,” which loads
Scene1 (our actual game), and “Exit,” which will exit the entire
application when the game is actually built (that is, it won’t actually
work within Unity).

Still need to add a title and music. Stretch goal: add a button to
toggle subtitles