VICTORY!
-got the ProtestCamera to move to the RIGHT POSITION AND ANGLE, using…
you guessed it… Lerp.

-a BIG lesson learned that was fucking everything up? turns out that
two cameras CANNOT EXIST IN THE EXACT SAME SPACE. They will move VERY
close to one another, within .0001 meters of one another, but the
coordinate will never be exactly the same. That was why the
ProtestCamera never actually switched, because I was originally
commanding it to only do so when it’s position was equal to the
MainCamera. Now I tell it to do so when the difference between its
position and the MainCamera is < .01, which is so close that no one
playing can tell. #hackz

-got rid of the special FirstEnemy shit because it doesn’t really work
as well now that the opening scene order has been rearranged

-put the MainCamera slightly closer to the player so the action feels a
bit more intimate

-the music is now attached to the ProtestSpawner game object… I think?
i could have sworn i put it there, but when I look in the inspector i
don’t see it. In any case, it certainly works the way it’s supposed to;
the music starts playing when the protesters appear

-one problem; couldn’t get the chant to start playing quite when it
should (it should start a few seconds earlier), but at this point, meh!
It’s already past 3:30 am. I’m getting in bed soon